fix: 투수기록 평균자책점 및 결과 계산
